+++ ecom/backend/database/migrations/2026_05_10_222557_create_orders_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('customer_email');
             $table->string('address');
             $table->integer('number_of_items')->default(0);
-            $table->float('total_amount')->default(0.0);
-            $table->float('total_cost')->default(0.0);
+            $table->decimal('total_amount', 10, 2)->default(0);
+            $table->decimal('total_cost', 10, 2)->default(0);
             $table->tinyInteger('status')->comment('0 => pending, 1 => shipped, 2 => delivered, 3 => canceled');
             $table->timestamps();
         });
